% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissolve_slices.R
\name{dissolve_slices}
\alias{dissolve_slices}
\title{dissolve adjacent tower slices into polygons or polygon centroid points}
\usage{
dissolve_slices(tower_sf, out_geometry = c("polygon", "centroid"),
  out_name = NULL)
}
\arguments{
\item{tower_sf}{sf data frame object containing columns `img_id` and `slice_id`
that uniquely identify each slice}

\item{out_geometry}{should output be dissolved polygons or the centroids of
the dissolved polygons}

\item{out_name}{optional name of file to save dissolved geomtery. must have file
extension writable by \code{\link[sf]{st_write}}}
}
\value{
sf data frame containing dissolved geometry of output type `out_geometry`
}
\description{
dissolve adjacent tower slices into polygons or polygon centroid points
}
