% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_jp2_file_df.R
\name{make_jp2_file_df}
\alias{make_jp2_file_df}
\title{Create data frame of jp2 and jp2 aux files from directory paths}
\usage{
make_jp2_file_df(jp2_dir, jp2_aux_dir)
}
\arguments{
\item{jp2_dir}{Path to directory containing .jp2 image files}

\item{jp2_aux_dir}{Path to directory containing .jp2.aux.xml files with
metadata about .jp2 image with corresponding name}
}
\description{
Create data frame of jp2 and jp2 aux files from directory paths
}
