% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_split_index.R
\name{make_split_index}
\alias{make_split_index}
\title{create integer sequence for splitting vectors into a list}
\usage{
make_split_index(vec_length, num_chunks)
}
\arguments{
\item{vec_length}{length of the output vector}

\item{num_chunks}{number of unique values in the output vector}
}
\description{
create integer sequence for splitting vectors into a list
}
