% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgb2grey.R
\name{rgb2grey}
\alias{rgb2grey}
\alias{rgb2gray}
\title{Convert RGB to greyscale using 'average' or 'gleam' methods}
\usage{
rgb2grey(my_brick, method = c("ave", "gleam"))
}
\arguments{
\item{my_brick}{Object of class RasterBrick or RasterStack from the raster package}

\item{method}{Method of converting RGB to greyscale}
}
\description{
Convert RGB to greyscale using 'average' or 'gleam' methods
}
\details{
For the 'gleam' method we assume that the RGB image has been gamma-compressed.
We then:
1. Gamma expand it based on the formula at the wikipedia link below
2. Calculate the arithmetic mean across the gamma-expanded channels
3. Rescale to 0:255

See PLoS One article referenced below.
}
\references{
\url{https://en.wikipedia.org/wiki/Grayscale#Converting_color_to_grayscale}
\url{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0029740}
}
