% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_jp2_image_dir.R
\name{score_jp2_image_dir}
\alias{score_jp2_image_dir}
\title{Score a directory of jp2 images with a Keras model}
\usage{
score_jp2_image_dir(jp2_dir, jp2_aux_dir, model_params_dput_file,
  model_h5_weights, score_outdir = NULL, compress_score_rds = FALSE,
  return_score = TRUE, verbose = FALSE, tile_overlap = 0,
  complete_image = FALSE)
}
\arguments{
\item{jp2_dir}{Path to directory containing .jp2 image files}

\item{jp2_aux_dir}{Path to directory containing .jp2.aux.xml files with
metadata about .jp2 image with corresponding name}

\item{model_params_dput_file}{Path to file containing \code{dput} export of parameters
used during training of the model contained in `model_h5_weights`.}

\item{model_h5_weights}{Saved h5 weights from a trained keras model, used to score
the image tiles.}

\item{score_outdir}{Path to directory where results of image scoring should be save}

\item{compress_score_rds}{Should saved score results be compressed}

\item{return_score}{Should list of score results be returned as an R object by the function}

\item{verbose}{Should messages about current step being processes be printed to screen?}

\item{tile_overlap}{Number of pixel overlap in adjacent tiles (in both X and Y directions).
See \code{\link{calc_slice_corners}}.}

\item{complete_image}{See \code{\link{calc_slice_corners}}.}
}
\value{
List of data frame with one row for each tile, containing:
- \code{jp2_path}
- \code{jp2_aux_path}
- Numeric tile ID number
- Tile corner cells in source jp2 pixel values
- List-column containing each tile's extent based on the source jp2's projected raster
- sf geometry column containing each tile's sf polygon
- List-column containing tile's RGB layers in a 4d array of dimension
\[1, \code{tile_n_cols}, \code{tile_n_rows}, 3\]
- Model predicted probability for each image tile

If score \code{outpath != NULL} each returned data frame will be saved to outpath as an
RDS file, compressed if \code{compress_score_rds == TRUE}.
}
\description{
Score a directory of jp2 images with a Keras model
}
