% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_slice_data_dir.R
\name{score_slice_data_dir}
\alias{score_slice_data_dir}
\title{Score a directory of tile data files with a Keras model}
\usage{
score_slice_data_dir(slice_data_dir, model_params_dput_file,
  model_h5_weights, score_outdir = NULL, compress_score_rds = FALSE,
  return_score = TRUE, keep_array = FALSE, verbose = FALSE)
}
\arguments{
\item{slice_data_dir}{Directory containing sliced image .rds files}

\item{model_params_dput_file}{Path to file containing \code{dput} export of parameters
used during training of the model contained in `model_h5_weights`.}

\item{model_h5_weights}{Saved h5 weights from a trained keras model, used to score
the image tiles.}

\item{score_outdir}{Path to directory where results of image scoring should be save}

\item{compress_score_rds}{Should saved score results be compressed}

\item{return_score}{Should list of score results be returned as an R object by the function}

\item{verbose}{Should messages about current step being processes be printed to screen?}
}
\value{
List of data frame with one row for each tile, containing:
List of lists containing model parameters and input `img_data` data frame with
additional column containing model predicted probabilities for each slice

If \code{score_outdir != NULL} each list will be saved to score_outdir as an
RDS file, compressed if \code{compress_score_rds == TRUE}.
}
\description{
Score a directory of tile data files with a Keras model
}
