% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_test_images.R
\name{score_test_images}
\alias{score_test_images}
\title{Score training test-set images}
\usage{
score_test_images(img_dir, model, prop_notower = 0.1, out_filename)
}
\arguments{
\item{img_dir}{Directory containing images to score}

\item{model}{Keras model to use for scoring}
}
\description{
Score training test-set images
}
