% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_tower_model.R
\name{train_tower_model}
\alias{train_tower_model}
\title{Train tower model with keras}
\usage{
train_tower_model(img_base_dir = "data/tiles/splits", img_dir = "orig",
  output_dir = "output", img_size = c(50, 50),
  img_horizontal_flip = TRUE, img_vertical_flip = TRUE,
  batch_size = 32, base_model = "vgg16", save_best_model_only = TRUE,
  add_small_final_layer = FALSE, small_layer_size = NULL,
  dense_structure = list(list(units = 256, dropout = 0.2), list(units =
  128, dropout = 0.2)), dense_optimizer = "rmsprop", dense_lr = 1e-05,
  dense_steps_per_epoch = 100, dense_epochs = 30,
  dense_validation_steps = 50, first_ft_unfreeze = "block4_conv1",
  first_ft_optimizer = "rmsprop", first_ft_lr = 1e-05,
  first_ft_steps_per_epoch = 100, first_ft_epochs = 30,
  first_ft_validation_steps = 50, do_second_ft = TRUE,
  second_ft_unfreeze = "block3_conv1", second_ft_optimizer = "rmsprop",
  second_ft_lr = 5e-06, second_ft_steps_per_epoch = 100,
  second_ft_epochs = 30, second_ft_validation_steps = 50,
  class_weights = NULL)
}
\description{
Train tower model with keras
}
